Extensible Enumerations
Added the option to reconcile enumeration option values caselessly via
the optional `caselessly` keyword argument.

Added the `.aliases` property to return a list of the enumeration
options which are aliases for the current enumeration option.

Added the `options()` method to return a `MappingProxy` for the
enumeration options to provide another way to access and iterate
through the options, and for parity with the Enumerific library’s
subclassed standard library `Enum` class, which also offers an
`options()` method that returns a `MappingProxy` for the enumeration
options.

Added a unit test for the `options()` method.

Updated the `README` file with information about the Enumerific
library’s `Enumeration` class and examples of its usage.