Extensible Enumerations
Added support for controlling whether an extensible Enumeration class
supports backfilling of options defined in subclasses or not through
the new `backfill` class definition keyword argument. By default
backfilling of options is prevented, but may be enabled by specifying
the `backfill` keyword argument and setting it to `True`.

Updated the `README` file to detail the new feature.

Added a unit test suite to ensure that default and backfilling behavior
work consistently as expected.