Extensible Enumerations
Added support for controlling the extensibility of extensible
Enumeration classes, which allows their extensibility via subclassing
or the registration of new options to be controlled if needed. This can
be configured when creating an extensible Enumeration class via the new
`extensible` class definition keyword argument.

Updated the `README` and `CHANGELOG` files to document the new
functionality and added unit testing to ensure that it works
consistently over time.